feat: mutable fee_vault and remove user feature

+++ programs/dynamic-fee-sharing/src/instructions/ix_claim_fee.rs
@@ -33,7 +33,8 @@ pub struct ClaimFeeCtx<'info> {
 
 pub fn handle_claim_fee(ctx: Context<ClaimFeeCtx>, index: u8) -> Result<()> {
     let mut fee_vault = ctx.accounts.fee_vault.load_mut()?;
-    let fee_being_claimed = fee_vault.validate_and_claim_fee(index, &ctx.accounts.user.key())?;
+    let fee_being_claimed =
+        fee_vault.validate_and_claim_fee(index as usize, &ctx.accounts.user.key())?;
 
     if fee_being_claimed > 0 {
         transfer_from_fee_vault(
